ci: build all images in one script
