+++ apps/deployment/package.json
@@ -3,9 +3,7 @@
   "type": "module",
   "devDependencies": {
     "@types/globby": "9.1.0",
-    "@types/minimist": "1.2.2",
     "globby": "13.1.3",
-    "minimist": "1.2.7",
     "zx": "7.1.1"
   }
 }
