+++ apps/deployment/src/data/admin-backend.json
@@ -1,5 +1,6 @@
 {
   "dockerfile_path": "apps/admin/backend/admin-backend.generated.Dockerfile",
+  "project_type": "rust",
   "apps": [
     {
       "EXECUTABLE_NAMES": ["admin-backend"],
