+++ apps/deployment/src/data/languages.json
@@ -1,5 +1,6 @@
 {
   "dockerfile_path": "apps/languages/${executable}.generated.Dockerfile",
+  "project_type": "rust",
   "apps": [
     {
       "EXECUTABLE_NAMES": ["rs-service"],
